refactor: move test case to relevant file
TestBillableLineItemsAcrossMonths moves to test_invoice_line_items.py, driving get_billable_line_items for the same coverage.
